mentor_agent_telegram/requirements.txt
flask>=3.0
langchain>=0.3
langchain-core>=0.3
langchain-community>=0.3
langchain-deepseek>=0.1
pymongo>=4.6
python-dotenv>=1.0
langchain-openai==1.1.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
python-telegram-bot>=20.0
apscheduler>=3.0